Change location of method
